WIP examples moved to plants
