+++ drake/systems/analysis/test/CMakeLists.txt
@@ -1,2 +1,2 @@
-drake_add_cc_test(simulator_test)
-target_link_libraries(simulator_test drakeSystemAnalysis)
+#drake_add_cc_test(simulator_test)
+#target_link_libraries(simulator_test drakeSystemAnalysis)
