+++ drake/systems/framework/CMakeLists.txt
@@ -7,8 +7,8 @@ set(sources
   diagram.cc
   diagram_builder.cc
   diagram_context.cc
-  examples/controlled_spring_mass_system.cc
-  examples/spring_mass_system.cc
+#  examples/controlled_spring_mass_system.cc
+#  examples/spring_mass_system.cc
   leaf_context.cc
   leaf_system.cc
   input_port_evaluator_interface.cc
