+++ drake/systems/plants/CMakeLists.txt
@@ -199,3 +199,7 @@ endif()
 
 add_subdirectory(rigid_body_plant)
 add_subdirectory(test)
+
+#System2.0 Plants
+add_subdirectory(spring_mass_system)
+add_subdirectory(controlled_spring_mass_system)
