+++ drake/systems/plants/controlled_spring_mass_system/controlled_spring_mass_system.cc
