+++ drake/systems/plants/spring_mass_system/spring_mass_system.h
