+++ drake/systems/plants/test/controlled_spring_mass_system_test.cc
