+++ drake/systems/plants/test/spring_mass_system_test.cc
